Rename json module to avoid naming conflicts
When running this package as `python -m sr.comp.http`, the 'http'
directory ends up at the start of the path. That means that anything
(including Flask) which tries to `import json` ends up trying to
import that module. I'm not sure whether this is expected as it seems
highly likely to cause issues for all sorts of packages, however
the simple solution is to rename our module to avoid the issue.